feat: Add Lexer for GraphQL AST w/ corresponding tests
Adding a lexer is the first step in being able to construct a parser
for GraphQL.AST. The lexical specification is documented ad-naseum
at http://facebook.github.io/graphql/draft/#sec-Appendix-Grammar-Summary.Lexical-Tokens

Alex was the lexer of choice due to its use in GHC. An alex specification
also serves as a formal specification for the grammar, making it easier to
modify as new RFCs are merged into the GraphQL specification. Monadic
parser combinators are used in other GraphQL APIs but these tend to
conflate both the lexing and parsing into a single step. While this is
convenient, the GraphQL specification gives both a lexical specification
and a BNF describing how a parser should be implemented, making it a great
fit for the alex and happy tools.