feat: Add Generators for GraphQL AST
Random generation of an abstract syntax tree (AST), coupled with
pretty printing, provide round trip testing of parsing and
pretty printing of the GraphQL AST. This commit provides custom
generators for each node in the GraphQL AST. In order to avoid
exponential blow up of AST size, we limit the maximum depth to size 2.
QuickCheck was used out of convenience, but moving to hedgehog wouldn't
be a bad idea if we want prettier error messages.